Mailbook
An simple to store email addresses. With two basic functions, Add and List
Nothing sophisitcated.

The framework used is Laravel 5.3 .
The application stores data in ./database/database.db, and sqlite database
file.
The source for writing and retrieving from Db is found in ./app/MB
directory. The directory contains the Contract file in Interfaces
directory, and the actual codes in Services directory.

Other codes can be found in the Controllers folder of the /app directory.

To run the sample.

type: php artisan serve
from the application root directory.